fix: [PL-48009]: add testcases
