+++ packages/uicore/src/components/FormikForm/__tests__/FormikForm.test.tsx
@@ -6,7 +6,7 @@
  */
 
 import React from 'react'
-import { act, findByText, fireEvent, queryByAttribute, render, screen, waitFor, within } from '@testing-library/react'
+import { act, findByText, fireEvent, render, screen, waitFor } from '@testing-library/react'
 import { Formik, FormikForm, FormInput } from '../FormikForm'
 import { Button } from '../../Button/Button'
 import { MultiTypeInputType, MultiTypeInputValue } from 'components/MultiTypeInput/MultiTypeInputUtils'
@@ -14,7 +14,7 @@ import { TooltipContextProvider } from '../../../frameworks/Tooltip/TooltipConte
 import userEvent from '@testing-library/user-event'
 import { getIdentifierFromName } from 'components/InputWithIdentifier/InputWithIdentifier'
 
-const renderFormikForm = (
+export const renderFormikForm = (
   fields: React.ReactNode,
   submitMethod = jest.fn(),
   initialValues = {},
@@ -598,84 +598,5 @@ describe('<FormInput.KVTagInput />', () => {
 
     expect(getIdentifierFromName('_-$33var-abc-1--', true)).toBe('_-$33var-abc-1--')
     expect(getIdentifierFromName('_-$33var-abc-1--', false)).toBe('_$33varabc1')
-  })
-})
-
-describe('Test InputWithIdentifier', () => {
-  function addNameAndAssertModifiedIdentifier(
-    name: string,
-    identifier: string,
-    input: HTMLInputElement,
-    idClass: HTMLInputElement
-  ) {
-    userEvent.type(input, name)
-    expect(input).toHaveValue(name)
-    expect(within(idClass).getByText(identifier)).toBeInTheDocument()
-    userEvent.clear(input)
-    expect(input).toHaveValue('')
-  }
-  test('allowHyphen false', async () => {
-    const onSubmit = jest.fn()
-    const { container } = render(
-      renderFormikForm(
-        <FormInput.InputWithIdentifier
-          inputName="name"
-          idName="identifier"
-          inputLabel="Name"
-          idLabel="ID"
-          isIdentifierEditable
-          maxInput={128}
-          useUnversialToolTipId
-          onIdentifierChangeCallback={onSubmit}
-          // not passing allowHyphen as its optional
-        />
-      )
-    )
-    const input = queryByAttribute('name', container, 'name') as HTMLInputElement
-    const idClass = container.querySelector('div.idValue') as HTMLInputElement
-
-    expect(input).toBeInTheDocument()
-    expect(idClass).toBeInTheDocument()
-
-    addNameAndAssertModifiedIdentifier('Test Name', 'Test_Name', input, idClass)
-    addNameAndAssertModifiedIdentifier('Test Name-123', 'Test_Name123', input, idClass)
-    addNameAndAssertModifiedIdentifier('Test Name-123-', 'Test_Name123', input, idClass)
-    addNameAndAssertModifiedIdentifier('-Test Name-123-', 'Test_Name123', input, idClass)
-    addNameAndAssertModifiedIdentifier('-$Test Name-123-', 'Test_Name123', input, idClass)
-    addNameAndAssertModifiedIdentifier('-$Test Name-123-', 'Test_Name123', input, idClass)
-    addNameAndAssertModifiedIdentifier('_-$33var1--', '_$33var1', input, idClass)
-    addNameAndAssertModifiedIdentifier('_-$33var-abc-1--', '_$33varabc1', input, idClass)
-  })
-  test('allowHyphen true', async () => {
-    const onSubmit = jest.fn()
-    const { container } = render(
-      renderFormikForm(
-        <FormInput.InputWithIdentifier
-          inputName="name"
-          idName="identifier"
-          inputLabel="Name"
-          idLabel="ID"
-          isIdentifierEditable
-          maxInput={128}
-          useUnversialToolTipId
-          onIdentifierChangeCallback={onSubmit}
-          allowHyphen={true}
-        />
-      )
-    )
-    const input = queryByAttribute('name', container, 'name') as HTMLInputElement
-    const idClass = container.querySelector('div.idValue') as HTMLInputElement
-
-    expect(input).toBeInTheDocument()
-    expect(idClass).toBeInTheDocument()
-
-    addNameAndAssertModifiedIdentifier('Test Name', 'Test_Name', input, idClass)
-    addNameAndAssertModifiedIdentifier('Test Name-123', 'Test_Name-123', input, idClass)
-    addNameAndAssertModifiedIdentifier('Test Name-123-', 'Test_Name-123-', input, idClass)
-    addNameAndAssertModifiedIdentifier('-Test Name-123-', 'Test_Name-123-', input, idClass)
-    addNameAndAssertModifiedIdentifier('-$Test Name-123-', 'Test_Name-123-', input, idClass)
-    addNameAndAssertModifiedIdentifier('-$Test Name-123-', 'Test_Name-123-', input, idClass)
-    addNameAndAssertModifiedIdentifier('_-$33var1--', '_-$33var1--', input, idClass)
-    addNameAndAssertModifiedIdentifier('_-$33var-abc-1--', '_-$33var-abc-1--', input, idClass)
   })
 })
